refactor(models): split multi-object files + add folder barrels
Both actions and weapons used to export multiple model objects from one
file, violating the one-entity-per-file rule in src/models/CLAUDE.md.

- src/models/common/actions.ts → actions/{combat,class,exploration}.ts
  plus actions/index.ts barrel re-exporting combatActions, classActions,
  explorationActions (and their types). Tests split to match.
- src/models/common/gear/weapons.ts → weapons/weapon-properties/
  weapon-mastery.ts. Added gear/index.ts barrel re-exporting armor,
  weapons, weaponProperties, weaponMastery, plus their types and
  Equipment. Tests split to match.

External consumers now import from the folder barrel (e.g. `import
{ Equipment, armor } from "src/models/common/gear"`). Existing deep
paths like src/models/common/gear/armor still work and are used by
siblings inside the gear folder.

Weapons keeps its {id}|{name} dual lookup for now — step 4 drops the
name variant.
