+++ src/models/common/gear/index.ts
@@ -1,0 +1,13 @@
+export { armor } from "./armor";
+export { weaponMastery } from "./weapon-mastery";
+export { weaponProperties } from "./weapon-properties";
+export { weapons } from "./weapons";
+
+export type { Armor, ArmorCategory } from "./armor";
+export type { Equipment } from "./equipment";
+export type {
+  WeaponMasteryDetails,
+  WeaponMasteryName,
+} from "./weapon-mastery";
+export type { WeaponProperty } from "./weapon-properties";
+export type { Weapon, WeaponProficiency, WeaponRange } from "./weapons";
