+++ src/models/common/gear/weapons.test.ts
@@ -1,4 +1,4 @@
-import { weaponMastery, weaponProperties, weapons } from "./weapons";
+import { weapons } from "./weapons";
 
 describe("weapons.get()", () => {
   it("returns a weapon by id", () => {
@@ -48,61 +48,5 @@ describe("weapons.find()", () => {
 describe("weapons.list()", () => {
   it("contains 38 weapons", () => {
     expect(weapons.list()).toHaveLength(38);
-  });
-});
-
-describe("weaponProperties.get()", () => {
-  it("returns a property by id", () => {
-    expect(weaponProperties.get({ id: "finesse" }).description).toBeTruthy();
-  });
-
-  it("throws on unknown id", () => {
-    expect(() => weaponProperties.get({ id: "glowing" })).toThrow(
-      /Unknown weapon property/,
-    );
-  });
-});
-
-describe("weaponProperties.find()", () => {
-  it("returns a property by id", () => {
-    expect(weaponProperties.find({ id: "finesse" })).toBeDefined();
-  });
-
-  it("returns undefined for unknown id", () => {
-    expect(weaponProperties.find({ id: "glowing" })).toBeUndefined();
-  });
-});
-
-describe("weaponProperties.list()", () => {
-  it("contains 10 properties", () => {
-    expect(weaponProperties.list()).toHaveLength(10);
-  });
-});
-
-describe("weaponMastery.get()", () => {
-  it("returns a mastery by id", () => {
-    expect(weaponMastery.get({ id: "cleave" }).description).toBeTruthy();
-  });
-
-  it("throws on unknown id", () => {
-    expect(() =>
-      weaponMastery.get({ id: "flurry" as unknown as "cleave" }),
-    ).toThrow(/Unknown weapon mastery/);
-  });
-});
-
-describe("weaponMastery.find()", () => {
-  it("returns a mastery by id", () => {
-    expect(weaponMastery.find({ id: "cleave" })).toBeDefined();
-  });
-
-  it("returns undefined for unknown id", () => {
-    expect(weaponMastery.find({ id: "flurry" })).toBeUndefined();
-  });
-});
-
-describe("weaponMastery.list()", () => {
-  it("contains 8 masteries", () => {
-    expect(weaponMastery.list()).toHaveLength(8);
   });
 });
